+++ changelog.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Notify in API: View all, create and delete
 - President in API: View all, create and delete
 - Reimbursement in API: View all, create and delete
+- Sponsorship in API: View all, create and delete
 
 ## [0.6.0] - 2020-12-07
 ### Added